Refactor datatable, alerts and toasts for deleting and errors

+++ web/src/lib/elements/ui/dataTable/data-table-actions.svelte
@@ -4,12 +4,15 @@
   import Button from '$lib/elements/ui/button/button.svelte';
   import { goto } from '$app/navigation';
   import { page } from '$app/stores';
+  import * as AlertDialog from '$lib/elements/ui/alert-dialog';
 
   export let id: string;
   export let deleteEntry: (id: string, additionalId?: string) => Promise<void>;
   export let getData: () => Promise<void>;
   export let editAvailable: boolean;
   export let additionalId: string;
+
+  let alertOpen: boolean = false;
 </script>
 
 <DropdownMenu.Root>
@@ -28,12 +31,28 @@
       <DropdownMenu.Item on:click={() => goto(`${$page.url}/${id}`)}>Editieren</DropdownMenu.Item>
     {/if}
     <DropdownMenu.Separator />
-    <DropdownMenu.Item
-      on:click={async () => {
-        await deleteEntry(id, additionalId);
-        await getData();
-      }}
-      class="text-red-600">Löschen</DropdownMenu.Item
-    >
+    <DropdownMenu.Item class="text-red-600" on:click={() => (alertOpen = true)}>Löschen</DropdownMenu.Item>
   </DropdownMenu.Content>
 </DropdownMenu.Root>
+
+<AlertDialog.Root bind:open={alertOpen}>
+  <AlertDialog.Trigger />
+  <AlertDialog.Content>
+    <AlertDialog.Header>
+      <AlertDialog.Title>Eintrag löschen?</AlertDialog.Title>
+      <AlertDialog.Description
+        >Sind Sie sicher, dass Sie diesen Eintrag löschen möchten? Diese Aktion kann nicht rückgängig gemacht werden.</AlertDialog.Description
+      >
+    </AlertDialog.Header>
+    <AlertDialog.Footer>
+      <AlertDialog.Cancel class="shadow-custom">Abbrechen</AlertDialog.Cancel>
+      <AlertDialog.Action
+        on:click={async () => {
+          await deleteEntry(id, additionalId);
+          await getData();
+        }}
+        class="text-red-600 shadow-custom">Löschen</AlertDialog.Action
+      >
+    </AlertDialog.Footer>
+  </AlertDialog.Content>
+</AlertDialog.Root>
